Server: Split Gap processing and stack bing up outside the main file.
Add documentation.
