+++ BLE_GattServer/main.cpp
@@ -1,3 +1,19 @@
+/* mbed Microcontroller Library
+ * Copyright (c) 2017 ARM Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 #include <stdio.h>
 
 #include "platform/Callback.h"
@@ -10,23 +26,16 @@
 #include "ble/GapAdvertisingParams.h"
 #include "ble/GapAdvertisingData.h"
 #include "ble/GattServer.h"
+#include "BLEProcess.h"
 
 using mbed::callback;
 
 /**
  * A Clock service that demonstrate the GattServer features.
  *
- * The clock service host three characteristics that respectively model the
- * current hour, minute and second. The value of the second characteristic is
+ * The clock service host three characteristics that model the current hour,
+ * minute and second of the clock. The value of the second characteristic is
  * incremented automatically by the system.
- *
- * When the value of the second characteristic reach 60, it is reset to 0 and
- * the value of the minute characteristic is incremented.
- *
- * Similarly when value of the minute characteristic reach 60, it is reset and
- * the value of the hour characteristic is incremented.
- *
- * Finaly when the value of the hour characteristic reach 24, it is reset to 0.
  *
  * A client can subscribe to updates of the clock characteristics and get
  * notified when one of the value is changed. Clients can also change value of
@@ -60,16 +69,19 @@ public:
         _second_char.setWriteAuthorizationCallback(this, &Self::authorize_client_write);
     }
 
-    void start(GattServer &server, events::EventQueue &event_queue)
+
+
+    void start(BLE &ble_interface, events::EventQueue &event_queue)
     {
-        if (_event_queue) {
+         if (_event_queue) {
             return;
         }
 
-        _server = &server;
+        _server = &ble_interface.gattServer();
         _event_queue = &event_queue;
 
         // register the service
+        printf("Adding demo service\r\n");
         ble_error_t err = _server->addService(_clock_service);
 
         if (err) {
@@ -99,14 +111,17 @@ public:
 
 private:
 
-//
-// Handlers called after a characteristic has been read, writen or notified
-//
+    /**
+     * Handler called when a notification or an indication has been sent.
+     */
     void when_data_sent(unsigned count)
     {
         printf("sent %u updates\r\n", count);
     }
 
+    /**
+     * Handler called after an attribute has been written.
+     */
     void when_data_written(const GattWriteCallbackParams *e)
     {
         printf("data written:\r\n");
@@ -133,6 +148,9 @@ private:
         printf("\r\n");
     }
 
+    /**
+     * Handler called after an attribute has been read.
+     */
     void when_data_read(const GattReadCallbackParams *e)
     {
         printf("data read:\r\n");
@@ -149,27 +167,44 @@ private:
         }
     }
 
-//
-// Client Characteristic Configuration Descriptors (CCCD) events handlers
-//
+    /**
+     * Handler called after a client has subscribed to notification or indication.
+     *
+     * @param handle Handle of the characteristic value affected by the change.
+     */
     void when_update_enabled(GattAttribute::Handle_t handle)
     {
         printf("update enabled on handle %d\r\n", handle);
     }
 
+    /**
+     * Handler called after a client has cancelled his subscription from
+     * notification or indication.
+     *
+     * @param handle Handle of the characteristic value affected by the change.
+     */
     void when_update_disabled(GattAttribute::Handle_t handle)
     {
         printf("update disabled on handle %d\r\n", handle);
     }
 
+    /**
+     * Handler called when an indication confirmation has been received.
+     *
+     * @param handle Handle of the characteristic value that has emitted the
+     * indication.
+     */
     void when_confirmation_received(GattAttribute::Handle_t handle)
     {
         printf("confirmation received on handle %d\r\n", handle);
     }
 
-//
-// Write Authorization handlers for the second, minute and hour characteristics
-//
+    /**
+     * Handler called when a write request is received.
+     *
+     * This handler verify that the value submitted by the client is valid before
+     * authorizing the operation.
+     */
     void authorize_client_write(GattWriteAuthCallbackParams *e)
     {
         printf("characteristic %u write authorization\r\n", e->handle);
@@ -196,9 +231,9 @@ private:
         e->authorizationReply = AUTH_CALLBACK_REPLY_SUCCESS;
     }
 
-//
-// Incrementation of seconds, minutes and hours characteristics
-//
+    /**
+     * Increment the second counter.
+     */
     void increment_second(void)
     {
         uint8_t second = 0;
@@ -221,6 +256,9 @@ private:
         }
     }
 
+    /**
+     * Increment the minute counter.
+     */
     void increment_minute(void)
     {
         uint8_t minute = 0;
@@ -243,6 +281,9 @@ private:
         }
     }
 
+    /**
+     * Increment the hour counter.
+     */
     void increment_hour(void)
     {
         uint8_t hour = 0;
@@ -351,175 +392,13 @@ private:
     events::EventQueue *_event_queue;
 };
 
-
-/**
- * Handle initialization adn shutdown of the BLE Instance.
- *
- * Setup advertising payload and manage advertising state.
- * Delegate to GattClientProcess once the connection is established.
- */
-class BLEProcess : private mbed::NonCopyable<BLEProcess> {
-public:
-    /**
-     * Construct a BLEProcess from an event queue and a ble interface.
-     *
-     * Call start() to initiate ble processing.
-     */
-    BLEProcess(events::EventQueue &event_queue, BLE &ble_interface, ClockService& service) :
-        _event_queue(event_queue),
-        _ble_interface(ble_interface),
-        _demonstration_service(service) {
-    }
-
-    ~BLEProcess()
-    {
-        stop();
-    }
-
-    /**
-     * Initialize the ble interface, configure it and start advertising.
-     */
-    bool start()
-    {
-        printf("Ble process started.\r\n");
-
-        if (_ble_interface.hasInitialized()) {
-            printf("Error: the ble instance has already been initialized.\r\n");
-            return false;
-        }
-
-        _ble_interface.onEventsToProcess(
-            makeFunctionPointer(this, &BLEProcess::schedule_ble_events)
-        );
-
-        ble_error_t error = _ble_interface.init(
-            this, &BLEProcess::when_init_complete
-        );
-
-        if (error) {
-            printf("Error: %u returned by BLE::init.\r\n", error);
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * Close existing connections and stop the process.
-     */
-    void stop()
-    {
-        if (_ble_interface.hasInitialized()) {
-            _ble_interface.shutdown();
-            printf("Ble process stopped.");
-        }
-    }
-
-private:
-
-    /**
-     * Schedule processing of events from the BLE middleware in the event queue.
-     */
-    void schedule_ble_events(BLE::OnEventsToProcessCallbackContext *event)
-    {
-        _event_queue.call(mbed::callback(&event->ble, &BLE::processEvents));
-    }
-
-    /**
-     * Sets up adverting payload and start advertising.
-     *
-     * This function is invoked when the ble interface is initialized.
-     */
-    void when_init_complete(BLE::InitializationCompleteCallbackContext *event)
-    {
-        if (event->error) {
-            printf("Error %u during the initialization\r\n", event->error);
-            return;
-        }
-        printf("Ble instance initialized\r\n");
-
-        printf("Adding demo service\r\n");
-        _demonstration_service.start(_ble_interface.gattServer(), _event_queue);
-
-        Gap &gap = _ble_interface.gap();
-        ble_error_t error = gap.setAdvertisingPayload(make_advertising_data());
-        if (error) {
-            printf("Error %u during gap.setAdvertisingPayload\r\n", error);
-            return;
-        }
-
-        gap.setAdvertisingParams(make_advertising_params());
-
-        gap.onConnection(this, &BLEProcess::when_connection);
-        gap.onDisconnection(this, &BLEProcess::when_disconnection);
-
-        start_advertising();
-    }
-
-    void when_connection(const Gap::ConnectionCallbackParams_t *connection_event)
-    {
-        printf("Connected.\r\n");
-    }
-
-    void when_disconnection(const Gap::DisconnectionCallbackParams_t *event)
-    {
-        printf("Disconnected.\r\n");
-        start_advertising();
-    }
-
-    void start_advertising(void)
-    {
-        ble_error_t error = _ble_interface.gap().startAdvertising();
-        if (error) {
-            printf("Error %u during gap.startAdvertising.\r\n", error);
-            return;
-        } else {
-            printf("Advertising started.\r\n");
-        }
-    }
-
-    static GapAdvertisingData make_advertising_data(void)
-    {
-        static const uint8_t device_name[] = "GattServer";
-        GapAdvertisingData advertising_data;
-
-        // add advertising flags
-        advertising_data.addFlags(
-            GapAdvertisingData::LE_GENERAL_DISCOVERABLE |
-            GapAdvertisingData::BREDR_NOT_SUPPORTED
-        );
-
-        // add device name
-        advertising_data.addData(
-            GapAdvertisingData::COMPLETE_LOCAL_NAME,
-            device_name,
-            sizeof(device_name)
-        );
-
-        return advertising_data;
-    }
-
-    static GapAdvertisingParams make_advertising_params(void)
-    {
-        return GapAdvertisingParams(
-            /* type */ GapAdvertisingParams::ADV_CONNECTABLE_UNDIRECTED,
-            /* interval */ GapAdvertisingParams::MSEC_TO_ADVERTISEMENT_DURATION_UNITS(500),
-            /* timeout */ 0
-        );
-    }
-
-    events::EventQueue &_event_queue;
-    BLE &_ble_interface;
-    ClockService &_demonstration_service;
-};
-
-
 int main() {
-
     BLE &ble_interface = BLE::Instance();
     events::EventQueue event_queue;
     ClockService demo_service;
-    BLEProcess ble_process(event_queue, ble_interface, demo_service);
+    BLEProcess ble_process(event_queue, ble_interface);
+
+    ble_process.on_init(callback(&demo_service, &ClockService::start));
 
     // bind the event queue to the ble interface, initialize the interface
     // and start advertising
